+++ lib/dead_end/display_invalid_blocks.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "banner"
 require_relative "capture_code_context"
 require_relative "display_code_with_line_numbers"
 
@@ -9,18 +8,13 @@ module DeadEnd
   class DisplayInvalidBlocks
     attr_reader :filename
 
-    def initialize(code_lines:, blocks:, io: $stderr, filename: nil, terminal: DEFAULT_VALUE, invalid_obj: WhoDisSyntaxError::Null.new)
-      @terminal = terminal == DEFAULT_VALUE ? io.isatty : terminal
-
-      @filename = filename
+    def initialize(code_lines:, blocks:, io: $stderr, filename: nil, terminal: DEFAULT_VALUE)
       @io = io
-
       @blocks = Array(blocks)
-
-      @invalid_lines = @blocks.map(&:lines).flatten
+      @filename = filename
       @code_lines = code_lines
 
-      @invalid_obj = invalid_obj
+      @terminal = terminal == DEFAULT_VALUE ? io.isatty : terminal
     end
 
     def document_ok?
@@ -30,46 +24,43 @@ module DeadEnd
     def call
       if document_ok?
         @io.puts "Syntax OK"
-      else
-        found_invalid_blocks
+        return self
       end
+
+      @io.puts("--> #{filename}") if filename
+      @io.puts
+      @blocks.each do |block|
+        display_block(block)
+      end
+
       self
     end
 
-    private def no_invalid_blocks
-      @io.puts <<~EOM
-      EOM
-    end
+    private def display_block(block)
+      lines = CaptureCodeContext.new(
+        blocks: block,
+        code_lines: @code_lines
+      ).call
 
-    private def found_invalid_blocks
-      @io.puts
-      if banner
-        @io.puts banner
-        @io.puts
+      document = DisplayCodeWithLineNumbers.new(
+        lines: lines,
+        terminal: @terminal,
+        highlight_lines: block.lines
+      ).call
+
+      RipperErrors.new(block.lines.map(&:original).join).call.errors.each do |e|
+        @io.puts e
       end
-      @io.puts("file: #{filename}") if filename
-      @io.puts <<~EOM
-        simplified:
+      @io.puts
 
-        #{indent(code_block)}
-      EOM
+      @io.puts(document)
     end
 
-    def banner
+    private def banner
       Banner.new(invalid_obj: @invalid_obj).call
     end
 
-    def indent(string, with: "    ")
-      string.each_line.map { |l| with + l }.join
-    end
-
-    def code_block
-      string = +""
-      string << code_with_context
-      string
-    end
-
-    def code_with_context
+    private def code_with_context
       lines = CaptureCodeContext.new(
         blocks: @blocks,
         code_lines: @code_lines
@@ -77,14 +68,6 @@ module DeadEnd
 
       DisplayCodeWithLineNumbers.new(
         lines: lines,
-        terminal: @terminal,
-        highlight_lines: @invalid_lines
-      ).call
-    end
-
-    def code_with_lines
-      DisplayCodeWithLineNumbers.new(
-        lines: @code_lines.select(&:visible?),
         terminal: @terminal,
         highlight_lines: @invalid_lines
       ).call