+++ spec/unit/code_search_spec.rb
@@ -42,8 +42,8 @@ module DeadEnd
       EOM
     end
 
-    it "handles no spaces between blocks" do
-      search = CodeSearch.new(<<~'EOM')
+    it "handles no spaces between blocks and trailing slash" do
+      source = <<~'EOM'
         require "rails_helper"
         RSpec.describe Foo, type: :model do
           describe "#bar" do
@@ -64,13 +64,14 @@ module DeadEnd
         end
       EOM
 
+      search = CodeSearch.new(source)
       search.call
 
       expect(search.invalid_blocks.join.strip).to eq('it "returns true" do # <== HERE')
     end
 
     it "handles no spaces between blocks" do
-      search = CodeSearch.new(<<~EOM)
+      source = <<~EOM
         context "foo bar" do
           it "bars the foo" do
             travel_to DateTime.new(2020, 10, 1, 10, 0, 0) do
@@ -81,13 +82,13 @@ module DeadEnd
           it "should" do
         end
       EOM
-
+      search = CodeSearch.new(source)
       search.call
 
       expect(search.invalid_blocks.join.strip).to eq('it "should" do')
     end
 
-    it "recording" do
+    it "records debugging steps to a directory" do
       Dir.mktmpdir do |dir|
         dir = Pathname(dir)
         search = CodeSearch.new(<<~EOM, record_dir: dir)
@@ -100,7 +101,7 @@ module DeadEnd
         search.call
 
         expect(search.record_dir.entries.map(&:to_s)).to include("1-add-1.txt")
-        expect(search.record_dir.join("1-add-1.txt").read).to eq(<<~EOM.indent(4))
+        expect(search.record_dir.join("1-add-1.txt").read).to eq(<<~EOM)
             1  class OH
             2    def hello
           ❯ 3    def hai
@@ -154,20 +155,15 @@ module DeadEnd
       it "finds hanging def in this project" do
         source_string = fixtures_dir.join("this_project_extra_def.rb.txt").read
         search = CodeSearch.new(source_string)
-
         search.call
 
-        blocks = search.invalid_blocks
-        io = StringIO.new
-        display = DisplayInvalidBlocks.new(
-          code_lines: search.code_lines,
-          blocks: blocks,
-          io: io
-        )
-        display.call
-        # puts io.string
+        document = DisplayCodeWithLineNumbers.new(
+          lines: search.code_lines.select(&:visible?),
+          terminal: false,
+          highlight_lines: search.invalid_blocks.flat_map(&:lines)
+        ).call
 
-        expect(display.code_with_lines.strip_control_codes).to include(<<~EOM)
+        expect(document).to include(<<~EOM)
           ❯ 36      def filename
         EOM
       end
@@ -208,18 +204,13 @@ module DeadEnd
         EOM
         search.call
 
-        blocks = search.invalid_blocks
-        io = StringIO.new
-        display = DisplayInvalidBlocks.new(
-          io: io,
-          blocks: blocks,
-          code_lines: search.code_lines,
-          filename: "fake/spec/lol.rb"
-        )
-        display.call
-        # io.string
+        document = DisplayCodeWithLineNumbers.new(
+          lines: search.code_lines.select(&:visible?),
+          terminal: false,
+          highlight_lines: search.invalid_blocks.flat_map(&:lines)
+        ).call
 
-        expect(display.code_with_lines).to include(<<~EOM)
+        expect(document).to include(<<~EOM)
              1  require 'rails_helper'
              2
              3  RSpec.describe AclassNameHere, type: :worker do