+++ spec/unit/display_invalid_blocks_spec.rb
@@ -16,6 +16,7 @@ module DeadEnd
 
       search = CodeSearch.new(syntax_string)
       search.call
+
       io = StringIO.new
       display = DisplayInvalidBlocks.new(
         io: io,
@@ -92,7 +93,6 @@ module DeadEnd
       )
       display.call
       expect(io.string).to include("❯ 2    def hello")
-      expect(io.string).to include("DeadEnd")
     end
 
     it " wraps code with github style codeblocks" do
@@ -107,12 +107,14 @@ module DeadEnd
 
       code_lines = CleanDocument.new(source: source).call.lines
       block = CodeBlock.new(lines: code_lines[1])
-      display = DisplayInvalidBlocks.new(
+      io = StringIO.new
+      DisplayInvalidBlocks.new(
+        io: io,
         blocks: block,
         terminal: false,
         code_lines: code_lines
-      )
-      expect(display.code_block).to eq(<<~EOM)
+      ).call
+      expect(io.string).to include(<<~EOM)
           1  class OH
         ❯ 2    def hello
           4    def hai
@@ -130,36 +132,36 @@ module DeadEnd
         end
       EOM
 
+      io = StringIO.new
       block = CodeBlock.new(lines: code_lines[1])
-      display = DisplayInvalidBlocks.new(
+      DisplayInvalidBlocks.new(
+        io: io,
         blocks: block,
         terminal: false,
         code_lines: code_lines
-      )
+      ).call
 
-      expect(display.code_with_lines).to eq(
-        [
-          "  1  class OH",
-          "❯ 2    def hello",
-          "  3    def hai",
-          "  4    end",
-          "  5  end",
-          ""
-        ].join($/)
-      )
+      expect(io.string).to include([
+        "  1  class OH",
+        "❯ 2    def hello",
+        "  4    end",
+        "  5  end",
+        ""
+      ].join($/))
 
       block = CodeBlock.new(lines: code_lines[1])
-      display = DisplayInvalidBlocks.new(
+      io = StringIO.new
+      DisplayInvalidBlocks.new(
+        io: io,
         blocks: block,
         terminal: true,
         code_lines: code_lines
-      )
+      ).call
 
-      expect(display.code_with_lines).to eq(
+      expect(io.string).to include(
         [
           "  1  class OH",
           ["❯ 2  ", DisplayCodeWithLineNumbers::TERMINAL_HIGHLIGHT, "  def hello"].join,
-          "  3    def hai",
           "  4    end",
           "  5  end",
           ""